工资账套管理、员工账套管理

+++ hr-service/src/main/java/com/zhengjian/hr/service/system/RoleService.java
@@ -40,6 +40,9 @@ public class RoleService {
      */
     @Transactional
     public boolean editRoleMenu(Integer rid, Integer[] mids) {
+        if (rid == null) {
+            return false;
+        }
         menuRoleMapper.deleteByRid(rid);
         if (mids == null || mids.length == 0) {
             return true;
